Revert "RDKAAMP-469 : Change sprintf to snprintf"
This reverts commit 02e353e2c63444a8189d90247ed0c84e7f90b301.

Revert "RDKAAMP-469 compilation warnings - sprintf use and implicit casting"

This reverts commit 907d422b5ce6899ada38be21e58ea62468381b0f.

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_22_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_22_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_22_1

+++ AampEvent.cpp
@@ -381,7 +381,7 @@ const std::vector<std::string> &MediaMetadataEvent::getLanguages() const
  */
 int MediaMetadataEvent::getLanguagesCount() const
 {
-	return (int)mLanguages.size();
+	return mLanguages.size();
 }
 
 /**
@@ -409,7 +409,7 @@ const std::vector<long> &MediaMetadataEvent::getBitrates() const
  */
 int MediaMetadataEvent::getBitratesCount() const
 {
-	return (int)mBitrates.size();
+	return mBitrates.size();
 }
 
 /**
@@ -467,7 +467,7 @@ const std::vector<float> &MediaMetadataEvent::getSupportedSpeeds() const
  */
 int MediaMetadataEvent::getSupportedSpeedCount() const
 {
-	return (int)mSupportedSpeeds.size();
+	return mSupportedSpeeds.size();
 }
 
 /**
@@ -793,7 +793,7 @@ const std::vector<float> &SupportedSpeedsChangedEvent::getSupportedSpeeds() cons
  */
 int SupportedSpeedsChangedEvent::getSupportedSpeedCount() const
 {
-	return (int)mSupportedSpeeds.size();
+	return mSupportedSpeeds.size();
 }
 
 /**
@@ -1280,7 +1280,7 @@ const std::vector<uint8_t> &ID3MetadataEvent::getMetadata() const
  */
 int ID3MetadataEvent::getMetadataSize() const
 {
-	return (int)mMetadata.size();
+	return mMetadata.size();
 }
 
 /**
